feat: crypto update celery
