+++ pyproject.toml
@@ -14,8 +14,8 @@ dependencies = [
     "django-redis (>=5.4.0,<6.0.0)",
     "requests (>=2.32.3,<3.0.0)",
     "gunicorn (>=23.0.0,<24.0.0)",
-    "django-apscheduler (>=0.7.0,<0.8.0)",
     "celery (>=5.5.2,<6.0.0)",
+    "django-celery-beat (>=2.8.0,<3.0.0)",
 ]
