+++ src/market/apps.py
@@ -4,7 +4,3 @@ from django.apps import AppConfig
 class MarketConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'market'
-
-    def ready(self):
-        from .scheduler.update_crypto import start_scheduler
-        start_scheduler()
